Move `getTestRunsUrlPath()` to base helpers
