+++ packages/base/src/helpers/__tests__/app.test.ts
@@ -1,4 +1,6 @@
-import {getCommonAppBaseURL} from '../app'
+import {getCommonAppBaseURL, getTestRunsUrlPath} from '../app'
+import {SpanTags} from '../interfaces'
+import {CI_JOB_URL, CI_PIPELINE_URL} from '../tags'
 
 describe('getCommonAppBaseUrl', () => {
   test('the base URL that is correct', () => {
@@ -31,5 +33,60 @@ describe('getCommonAppBaseUrl', () => {
     expect(getCommonAppBaseURL('us5.datadoghq.com', '')).toBe('https://us5.datadoghq.com/')
     expect(getCommonAppBaseURL('us5.datadoghq.com', 'app')).toBe('https://us5.datadoghq.com/')
     expect(getCommonAppBaseURL('us5.datadoghq.com', 'myorg')).toBe('https://myorg.datadoghq.com/')
+  })
+})
+
+describe('getTestRunsUrlPath', () => {
+  it('should return empty string when no CI_JOB_URL or CI_PIPELINE_URL', () => {
+    const spanTags: SpanTags = {}
+    const result = getTestRunsUrlPath(spanTags)
+    expect(result).toBe('')
+  })
+
+  it('should return path with job URL when CI_JOB_URL is present', () => {
+    const spanTags: SpanTags = {
+      [CI_JOB_URL]: 'https://ci.example.com/job/123',
+    }
+    const result = getTestRunsUrlPath(spanTags)
+    expect(result).toBe('ci/test-runs?query=%20%40ci.job.url%3A%22https%3A%2F%2Fci.example.com%2Fjob%2F123%22')
+  })
+
+  it('should return path with pipeline URL when CI_PIPELINE_URL is present but no CI_JOB_URL', () => {
+    const spanTags: SpanTags = {
+      [CI_PIPELINE_URL]: 'https://ci.example.com/pipeline/456',
+    }
+    const result = getTestRunsUrlPath(spanTags)
+    expect(result).toBe(
+      'ci/test-runs?query=%20%40ci.pipeline.url%3A%22https%3A%2F%2Fci.example.com%2Fpipeline%2F456%22'
+    )
+  })
+
+  it('should prefer CI_JOB_URL over CI_PIPELINE_URL when both are present', () => {
+    const spanTags: SpanTags = {
+      [CI_JOB_URL]: 'https://ci.example.com/job/123',
+      [CI_PIPELINE_URL]: 'https://ci.example.com/pipeline/456',
+    }
+    const result = getTestRunsUrlPath(spanTags)
+    expect(result).toBe('ci/test-runs?query=%20%40ci.job.url%3A%22https%3A%2F%2Fci.example.com%2Fjob%2F123%22')
+  })
+
+  it('should include queryPrefix when provided', () => {
+    const spanTags: SpanTags = {
+      [CI_JOB_URL]: 'https://ci.example.com/job/123',
+    }
+    const result = getTestRunsUrlPath(spanTags, '@service:my-service')
+    expect(result).toBe(
+      'ci/test-runs?query=%40service%3Amy-service%20%40ci.job.url%3A%22https%3A%2F%2Fci.example.com%2Fjob%2F123%22'
+    )
+  })
+
+  it('should handle special characters in URLs', () => {
+    const spanTags: SpanTags = {
+      [CI_JOB_URL]: 'https://ci.example.com/job/my-job?param=value&other=123',
+    }
+    const result = getTestRunsUrlPath(spanTags)
+    expect(result).toBe(
+      'ci/test-runs?query=%20%40ci.job.url%3A%22https%3A%2F%2Fci.example.com%2Fjob%2Fmy-job%3Fparam%3Dvalue%26other%3D123%22'
+    )
   })
 })
