+++ packages/datadog-ci/src/commands/junit/utils.ts
@@ -1,27 +1,6 @@
-import {getBaseUrl} from '@datadog/datadog-ci-base/helpers/app'
+import {getBaseUrl, getTestRunsUrlPath} from '@datadog/datadog-ci-base/helpers/app'
 import {SpanTags} from '@datadog/datadog-ci-base/helpers/interfaces'
-import {
-  CI_JOB_URL,
-  CI_PIPELINE_URL,
-  GIT_BRANCH,
-  GIT_REPOSITORY_URL,
-  GIT_SHA,
-} from '@datadog/datadog-ci-base/helpers/tags'
-
-export const getTestRunsUrlPath = (spanTags: SpanTags, queryPrefix = ''): string => {
-  if (!spanTags[CI_PIPELINE_URL] && !spanTags[CI_JOB_URL]) {
-    return ''
-  }
-
-  let query = queryPrefix
-  if (spanTags[CI_JOB_URL]) {
-    query += ` @ci.job.url:"${spanTags[CI_JOB_URL]}"`
-  } else if (spanTags[CI_PIPELINE_URL]) {
-    query += ` @ci.pipeline.url:"${spanTags[CI_PIPELINE_URL]}"`
-  }
-
-  return `ci/test-runs?query=${encodeURIComponent(query)}`
-}
+import {GIT_BRANCH, GIT_REPOSITORY_URL, GIT_SHA} from '@datadog/datadog-ci-base/helpers/tags'
 
 export const getTestRunsUrl = (spanTags: SpanTags, queryPrefix = ''): string => {
   const path = getTestRunsUrlPath(spanTags, queryPrefix)
