+++ packages/plugin-synthetics/src/reporters/default.ts
@@ -2,11 +2,10 @@ import type {TunnelReporter} from '../tunnel/tunnel'
 import type {CommandContext} from '@datadog/datadog-ci-base/helpers/interfaces'
 import type {Writable} from 'stream'
 
+import {getTestRunsUrlPath} from '@datadog/datadog-ci-base/helpers/app'
 import {CI_JOB_URL, CI_PIPELINE_URL} from '@datadog/datadog-ci-base/helpers/tags'
 import chalk from 'chalk'
 import ora from 'ora'
-
-import {getTestRunsUrlPath} from '../../junit/utils'
 
 import {
   Assertion,
